+++ core/src/com/mygdx/gameobjects/Mine.java
@@ -1,6 +1,6 @@
 package com.mygdx.gameobjects;
 
-public class Mine extends Building{
+public class Mine extends Building {
 	int coinConstant=25;
 	public Mine(){
 		
@@ -8,8 +8,8 @@ public class Mine extends Building{
 	public Mine(float x, float y, int width, int height ,long producedTime,int idGroup, String id){
 		super(x,y,width,height,producedTime,idGroup,id,10);
 	}
-	public int extract(long time) {
-		long timeDifference=time-storedTime;
+	public int extract(double time) {
+		double timeDifference=time-storedTime;
 		if(timeDifference>produceTime){
 			storedTime=time;
 			return coinConstant;
