+++ core/src/com/mygdx/gameobjects/iFire.java
@@ -4,6 +4,6 @@ import java.util.List;
 import java.util.UUID;
 
 public interface iFire {
-	public GameObject fire(long time,List<? extends GameObject> objectList);		
+	public GameObject fire(double time,List<? extends GameObject> objectList);		
 	public GameObject scanForTarget(List<? extends GameObject> objectList);	
 }
